API standard integer type: int -> long throughout
The use of int as the standard type for API parameters caused problems
upstream. The standard API integer is now long, which is 32 bits on
32 bit machines and 64 bits on 64 bit machines.

- All six API headers (terminal, graphics, sound, services, network,
  option): parameters, return types, callback/override typedefs, and
  API struct fields (event, menu, sequencer, option records) are long.
- All implementations updated: linux, windows, macosx (including the
  cocoa bridge), stub, portable (gnome_widgets, managerc), utils, and
  the C++ wrappers. OS interop (X11, ALSA, POSIX, Win32, Cocoa) keeps
  system types with conversion at call boundaries.
- API full scale values move INT_MAX -> LONG_MAX (angles, sound
  levels, latitude/longitude/altitude, color components, joystick
  axes, scroll positions). Scale constants are written relative to
  LONG_MAX (e.g. MIDI divisor LONG_MAX/128+1, RGB LONG_MAX/256+1) so
  they are correct for either long width, and scaling math divides
  first so it cannot overflow at full scale.
- All tests, games, and demo programs converted, including printf
  formats to %ld (audited by hand; the bypass stdio hides format
  mismatches from gcc).
- Fixes made in passing: joystick deadband/delta overflow at long
  scale, abs->labs truncations, ami_time post-2038 truncation,
  ami_elapsed wrap constant (used the true 48h clock modulus),
  windows MIDI fine controller && vs & typos, macosx joystick axis
  rescale at the bridge boundary, stale stub function names, and
  tests/hangc.c stale pre-rename API names.

Verified: full linux make all clean (zero warnings); services, stdio,
and network loopback tests pass at runtime; windows sources cross
compile clean under mingw-w64. Untested: windows/macosx runtime and
32 bit builds.

+++ graph_games/backgammon.c
@@ -44,7 +44,7 @@ Color helper
 
 *******************************************************************************/
 
-#define CLR(v) ((v) * (INT_MAX / 255))
+#define CLR(v) ((v) * (LONG_MAX / 255))
 
 /*******************************************************************************
 
@@ -401,8 +401,8 @@ void show_message(const char *msg)
 void play_sound(int note, int dur)
 {
     if (!sound_enabled) return;
-    ami_noteon(AMI_SYNTH_OUT, 0, 1, note, INT_MAX);
-    ami_noteoff(AMI_SYNTH_OUT, ami_curtimeout()+dur, 1, note, INT_MAX);
+    ami_noteon(AMI_SYNTH_OUT, 0, 1, note, LONG_MAX);
+    ami_noteoff(AMI_SYNTH_OUT, ami_curtimeout()+dur, 1, note, LONG_MAX);
 }
 
 #define WAVE_SLIDE 1
@@ -2276,14 +2276,14 @@ int main(void)
         FILE *pf;
         pf = fopen("backgammon.pos", "r");
         if (pf) {
-            int pw, ph, ppx, ppy;
+            long pw, ph, ppx, ppy;
             char line[256];
 
             /* skip header line */
             if (fgets(line, sizeof(line), pf) &&
                 fgets(line, sizeof(line), pf) && /* blank line */
                 fgets(line, sizeof(line), pf)) {
-                if (sscanf(line, "width: %d height: %d position x: %d position y: %d",
+                if (sscanf(line, "width: %ld height: %ld position x: %ld position y: %ld",
                            &pw, &ph, &ppx, &ppy) >= 2) {
                     if (pw > 100 && ph > 100) {
                         ami_setsizg(stdout, pw, ph);
@@ -2501,13 +2501,13 @@ int main(void)
     /* save window position/size */
     {
         FILE *pf;
-        int wx, wy;
+        long wx, wy;
 
         ami_getsizg(stdout, &wx, &wy);
         pf = fopen("backgammon.pos", "w");
         if (pf) {
             fprintf(pf, "Backgammon position\n\n");
-            fprintf(pf, "width: %d height: %d position x: 0 position y: 0\n",
+            fprintf(pf, "width: %ld height: %ld position x: 0 position y: 0\n",
                     wx, wy);
             fclose(pf);
         }
