+++ sound_programs/connectwave.c
@@ -40,11 +40,11 @@ The result returned is the sample size in bytes.
 
 *******************************************************************************/
 
-int splsiz(int chan, int bits)
+long splsiz(long chan, long bits)
 
 {
 
-    int size;
+    long size;
 
     size = bits/8; /* find number of bytes basic */
     if (bits%8) size++; /* round up to nearest byte */
@@ -57,10 +57,10 @@ int main(int argc, char **argv)
 
 {
 
-    int sport;         /* source port */
-    int dport;         /* destination port */
-    int ssize;         /* sample size in bytes */
-    int sbuf;          /* number of samples per buffer */
+    long sport;        /* source port */
+    long dport;        /* destination port */
+    long ssize;        /* sample size in bytes */
+    long sbuf;         /* number of samples per buffer */
     byte buff[BUFLEN]; /* buffer for samples */
 
     if (argc != 1 && argc != 3) {
